requirements.txt
Flask
Flask-Cors
openai
python-dotenv
psycopg2-binary
